+++ src/main/java/joshie/harvest/buildings/BuildingStage.java
@@ -2,7 +2,9 @@ package joshie.harvest.buildings;
 
 import joshie.harvest.buildings.placeable.Placeable;
 import joshie.harvest.buildings.placeable.Placeable.ConstructionStage;
+import joshie.harvest.core.handlers.HFTrackers;
 import joshie.harvest.core.util.Direction;
+import joshie.harvest.town.TownDataServer;
 import joshie.harvest.town.TownHelper;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.Mirror;
@@ -17,6 +19,7 @@ public class BuildingStage {
     public BuildingImpl building;
     public Direction direction;
     public ConstructionStage stage;
+    public boolean basement;
     public int index;
     public BlockPos pos;
 
@@ -26,6 +29,7 @@ public class BuildingStage {
         this.direction = Direction.withMirrorAndRotation(mirror, rotation);
         this.stage = ConstructionStage.BUILD;
         this.index = 0;
+        this.basement = true;
         this.pos = pos.add(0, building.getOffsetY(), 0);
     }
 
@@ -33,12 +37,20 @@ public class BuildingStage {
         return index < building.getFullList().size() ? building.getFullList().get(index): null;
     }
 
+    public BuildingImpl getBuilding() {
+        return building;
+    }
+
+    public ConstructionStage getStage() {
+        return stage;
+    }
+
     public BlockPos getPos(Placeable placeable) {
         return placeable.getTransformedPosition(pos, direction);
     }
 
     public double getDistance(Placeable placeable) {
-        return placeable.getOffsetPos().getY() <= -building.getOffsetY()? 256D: 64D;
+        return placeable.getOffsetPos().getY() <= -building.getOffsetY()? 256D: 96D;
     }
 
     private boolean increaseIndex(World world) {
@@ -60,6 +72,19 @@ public class BuildingStage {
 
                 TownHelper.getClosestTownToBlockPos(world, pos).addBuilding(world, building, direction, pos);
                 return true;
+            }
+
+            basement = true; //Reset the basement
+            TownHelper.<TownDataServer>getClosestTownToBlockPos(world, pos).syncBuildings(world);
+        }
+
+        //Resync when we change
+        if (basement) {
+            Placeable placeable = next();
+            if (placeable.getOffsetPos().getY() > -getBuilding().getOffsetY()) {
+                basement = false;
+                HFTrackers.markDirty(world);
+                TownHelper.<TownDataServer>getClosestTownToBlockPos(world, pos).syncBuildings(world);
             }
         }
 
@@ -107,6 +132,7 @@ public class BuildingStage {
         stage.building = BuildingRegistry.REGISTRY.getValue(new ResourceLocation(nbt.getString("CurrentlyBuilding")));
         stage.direction = Direction.valueOf(nbt.getString("Direction"));
         stage.pos = new BlockPos(nbt.getInteger("BuildingX"), nbt.getInteger("BuildingY"), nbt.getInteger("BuildingZ"));
+        stage.basement = nbt.getBoolean("Basement");
 
         if (nbt.hasKey("Stage")) {
             stage.index = nbt.getInteger("Index");
@@ -122,6 +148,7 @@ public class BuildingStage {
         nbt.setInteger("BuildingX", pos.getX());
         nbt.setInteger("BuildingY", pos.getY());
         nbt.setInteger("BuildingZ", pos.getZ());
+        nbt.setBoolean("Basement", basement);
 
         if (stage != null) {
             nbt.setInteger("Stage", stage.ordinal());