+++ src/main/java/joshie/harvest/npc/greeting/GreetingMultiple.java
@@ -1,6 +1,9 @@
 package joshie.harvest.npc.greeting;
 
 import joshie.harvest.api.npc.IConditionalGreeting;
+import joshie.harvest.api.npc.INPC;
+import net.minecraft.entity.EntityAgeable;
+import net.minecraft.entity.player.EntityPlayer;
 
 import static joshie.harvest.core.lib.HFModInfo.MODID;
 
@@ -17,7 +20,7 @@ public class GreetingMultiple implements IConditionalGreeting {
     }
 
     @Override
-    public String getUnlocalizedText() {
+    public String getUnlocalizedText(EntityPlayer player, EntityAgeable ageable, INPC npc) {
         return text;
     }
 }