+++ src/main/java/joshie/harvest/town/TownDataClient.java
@@ -1,3 +1,11 @@
 package joshie.harvest.town;
 
-public class TownDataClient extends TownData {}
+import joshie.harvest.buildings.BuildingStage;
+
+import java.util.LinkedList;
+
+public class TownDataClient extends TownData {
+    public void setBuilding(LinkedList<BuildingStage> building) {
+        this.building = building;
+    }
+}